prefer distro package of psycopg2
It's simpler to use the distro package for psycopg2 since compiling it
requires headers for Python 3 and Postgres. This way you will also get
all the niceties that distro packages give you like proper debug symbols
and compiler flags.

For the other dependencies it's not much of an issue as they are pure
Python.